# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/account:
    delete:
      tags: [Auth]
      operationId: deleteCurrentAccount
      summary: Permanently delete the signed-in account and its cloud-synced data
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Account and cloud-synced data were deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionSuccess'
        '401':
          description: Sign-in is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/data:
    delete:
      tags: [Auth]
      operationId: deleteCurrentData
      summary: Delete selected personal cloud data without deleting the account
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataDeletionInput'
      responses:
        '200':
          description: Selected personal cloud data were deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataDeletionResponse'
        '400':
          description: The requested deletion scope is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Sign-in is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema: { type: string }
        - name: state
          in: query
          required: false
          schema: { type: string }
        - name: iss
          in: query
          required: false
          schema: { type: string, format: uri }
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
...
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
...
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id: { type: string }
        email: { type: [string, 'null'], format: email }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        profileImageUrl: { type: [string, 'null'] }
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    AccountDeletionSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
    DataDeletionInput:
      type: object
      required: [scope]
      properties:
        scope:
          type: string
          enum: [route_data, private_projects, all_cloud_data]
    DataDeletionResponse:
      type: object
      required: [success, scope]
      properties:
        success: { type: boolean, const: true }
        scope:
          type: string
          enum: [route_data, private_projects, all_cloud_data]
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code: { type: string, minLength: 1 }
        code_verifier: { type: string, minLength: 1 }
        redirect_uri: { type: string, minLength: 1, format: uri }
        state: { type: string, minLength: 1 }
        nonce: { type: string, minLength: 1 }
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token: { type: string }
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...